feat: implement user authentication with login, logout, and role-based access control

+++ src/routes/artists/[id]/+page.server.ts
@@ -1,12 +1,16 @@
 import type { PageServerLoad } from './$types';
-import apiClient from '$lib/server/api-clients.server';
+import apiClient, { withAuth } from '$lib/server/api-clients.server';
 import { fail, redirect } from '@sveltejs/kit';
 import type { Actions} from './$types';
 
-export const load: PageServerLoad = async ({ params }) => {
+export const load: PageServerLoad = async ({ params, cookies }) => {
     const id = params.id; 
+    const token = cookies.get('token');
+    if (!token) {
+        throw redirect(303, '/login');
+    }
     try {
-        const response = await apiClient.get(`/artists/${id}`);
+        const response = await apiClient.get(`/artists/${id}`, withAuth(token));
         let artist = null;
         if (response.status === 200) {
             artist = response.data.data;
@@ -28,11 +32,16 @@ export const load: PageServerLoad = async ({ params }) => {
 };
 
 export const actions: Actions = {
-    delete: async ({ params }) => {
+    delete: async ({ params, cookies }) => {
         const { id } = params;
+        const token = cookies.get('token');
+
+        if (!token) {
+            throw redirect(303, '/login');
+        }
 
         try {
-            await apiClient.delete(`/artists/${id}`);
+            await apiClient.delete(`/artists/${id}`, withAuth(token));
         } catch (err) {
             console.error(err);
             return fail(500, { message: 'ลบไม่สำเร็จ' });
